Refactor repository into a new 'em-app' package
This refactoring creates a new, installable Python package named 'em-app' for electromagnetic applications. The package uses 'MTFLibrary' as a dependency.

The changes include:
- Restructuring the source code into the `src/em_app` directory.
- Updating `pyproject.toml` to define the new package, its metadata, and its dependency on `mtflib`.
- Removing the old `setup.py` from `MTFLibrary`, which is no longer needed.
- Reorganizing the tests to match the new package structure and updating their imports.
