+++ src/em_app/biot_savart.py
