+++ src/em_app/plotting.py
