+++ src/em_app/rectangular_loop.py
