+++ src/em_app/straight_wire.py
